Tests for L acceptance
L can be used in the DOM or DOW field.
In DOM it means the last day of the month, with an offset if set (to second-last, ... day
In DOW it means either Saturday or the last x-day of the month, e. g. 2L means the last Monday of the month.