Remove all dependencies except typescript
To try to reduce the node_modules and dependencies to a minimum, I figured I could rely on simply running the tests
with the node built in test runner. This means that aside from typescript, we have no dependency.

One upside of this method is that we use the built version of the lib and do not need any typescript processor to
run the tests. This is fine because :
- this project is very small and simple
- the tests do not need to be typed
- the build step is quick enough to be built before each test run

This also moves the file in the root. Since there are only two files, why bother with a src directory?

It also gets rid of the linter. This is because :
- I don't think tslint is supported anymore
- I already have a default prettier running in my text editor
- I didn't feel like adding eslint in the project with some configs
- The project is small so manual linting is still doable